refactor(gauge value): Refactored gauge value calculations and separator rendering in system...
# Project State

## Current Focus
Refactored gauge value calculations and separator rendering in system monitor and widget gallery examples

## Context
The changes improve type consistency and fix potential overflow issues in separator rendering calculations

## Completed
- [x] Removed unnecessary `as f64` casts in gauge value calculations
- [x] Fixed potential overflow in separator rendering by using consistent type conversions
- [x] Standardized separator rendering logic between system monitor and widget gallery

## In Progress
- [x] No active work in progress

## Blockers
- None identified

## Next Steps
1. Verify no visual regressions in affected examples
2. Consider adding unit tests for separator rendering logic
---
category: refactor
scope: gauge value
---

+++ examples/_apps/system_monitor.rs
@@ -329,8 +329,8 @@ impl SystemMonitor {
     }
 
     fn update_gauges(&mut self) {
-        let cpu = self.data.cpu_hist.current() as f64;
-        let mem = self.data.mem_hist.current() as f64;
+        let cpu = self.data.cpu_hist.current();
+        let mem = self.data.mem_hist.current();
         self.cpu_gauge.set_value(cpu);
         self.mem_gauge.set_value(mem);
 
@@ -372,7 +372,7 @@ impl Widget for SystemMonitor {
 
         // Separator
         for x in 0..area.width {
-            let idx = (area.width + x as u16) as usize;
+            let idx = (area.width + x) as usize;
             if idx < plane.cells.len() { plane.cells[idx].char = '─'; plane.cells[idx].fg = t.outline; }
         }
